move ObjectModel into file
